+++ src/pipeline/gui.py
@@ -2,7 +2,7 @@
 from __future__ import annotations # Delays annotation evaluation, allowing modern 3.10+ type syntax and forward references in older Python versions 3.8 and 3.9
 import threading
 
-from pipeline.interface import gui_eds
+from pipeline.server.trend_server_eds import launch_server_for_web_gui_eds_trend_specific 
 
 GLOBAL_SHUTDOWN_EVENT = threading.Event()
 
@@ -18,4 +18,4 @@ import signal
 #signal.signal(signal.SIGINT, handle_interrupt)
 
 
-gui_eds.main()
+launch_server_for_web_gui_eds_trend_specific ()